+++ client/main.html
@@ -48,21 +48,6 @@
         </nav>
 
     </div>
-    <!--3. 장르-->
-    <!-- <section class="contents__genre">
-        <div class="genrebar">
-            <button type="button" id="genrebar__home">홈</button>
-            <button type="button" id="genrebar__days">요일연재</button>
-            <button type="button" id="genrebar__webtoon">웹툰</button>
-            <button type="button" id="genrebar__boys">소년</button>
-            <button type="button" id="genrebar__drama">드라마</button>
-            <button type="button" id="genrebar__romance">로맨스</button>
-            <button type="button" id="genrebar__rofan">로판</button>
-            <button type="button" id="genrebar__action">액션무협</button>
-            <button type="button" id="genrebar__bl">BL</button>
-        </div>     
-    
-    </section> -->
     <section class="contents__genre">
         <div class="genrebar">
             <button type="button" class="genrebar__tab">홈</button>
@@ -194,8 +179,7 @@
     </ul>
 </section>
 <script type="module" src="main.js"></script>
-<script type="module" src="innerHtml.js"></script>
-<script type="module" src="w-contents.js"></script>
-<script type="module" src="data_cover.js"></script>
+<script type="module" src="./module/w-contents.js"></script>
+<script type="module" src="./data/data_cover.js"></script>
 </body>
 </html>
